fix rubocop error

+++ app/helpers/posts_helper.rb
@@ -3,31 +3,19 @@ module PostsHelper
     if user_signed_in?
       post.user.name
     else
-      "John Doe:"
+      'John Doe:'
     end
   end
 
   def sign_in_user
-    if !user_signed_in?
-      "Sign Up or Sign In to see who posted!!!"    
-    end
+    'Sign Up or Sign In to see who posted!!!' unless user_signed_in?
   end
 
-def is_current_user(post)
-  if current_user == post.user
-    
-    link_to 'Edit', edit_post_path(post) 
-    
+  def is_current_user(post) # rubocop:disable Naming/PredicateName
+    link_to 'Edit', edit_post_path(post) if current_user == post.user
   end
- 
-end
 
-def destroy_post(post)
-  if current_user == post.user  
-    
-    link_to 'Destroy', post, method: :delete, data: { confirm: 'Are you sure?' }
+  def destroy_post(post)
+    link_to 'Destroy', post, method: :delete, data: { confirm: 'Are you sure?' } if current_user == post.user
   end
-end
-
-
 end
